% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizers.R
\name{findOverlapsFast}
\alias{findOverlapsFast}
\title{A fast, Rust-based findOverlaps analog.}
\usage{
findOverlapsFast(query, subject, ...)
}
\arguments{
\item{query}{A GRanges object}

\item{subject}{A Tokenizer object (your Rust-based subject)}

\item{...}{Further arguments for consistency with findOverlaps signature}
}
\value{
A Hits object containing the query/subject index matches.
}
\description{
A fast, Rust-based findOverlaps analog.
}
