% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{igd_search}
\alias{igd_search}
\title{Search igd with a bed file}
\usage{
igd_search(database_path, query_path)
}
\arguments{
\item{database_path}{A string representing the path to the database igd file.}

\item{query_path}{A string representing the path to the query bed file.}
}
\description{
Search igd with a bed file
}
