% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{r_md5_digest}
\alias{r_md5_digest}
\title{create md5_digest}
\usage{
r_md5_digest(readable)
}
\arguments{
\item{readable}{a readable string}
}
\value{
an md5 digest on success
}
\description{
create an md5 digest from a readable string
}
\examples{
\dontrun{
r_md5_digest('ATCG')
}

}
