% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_tokens_from_gtok}
\alias{read_tokens_from_gtok}
\title{Write tokens to a gtok file}
\usage{
read_tokens_from_gtok(filename)
}
\arguments{
\item{filename}{A string representing the path to the gtok file.}
}
\description{
Write tokens to a gtok file
}
