% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_igd_search}
\alias{r_igd_search}
\title{RUST WRAPPERS SHOULD BE MINIMAL. HANDLE DATA STRUCTURES IN IGD
Search igd with a bed file}
\usage{
r_igd_search(database_path, query_path)
}
\arguments{
\item{database_path}{A string representing the path to the database igd file.}

\item{query_path}{A string representing the path to the query bed file.}
}
\description{
RUST WRAPPERS SHOULD BE MINIMAL. HANDLE DATA STRUCTURES IN IGD
Search igd with a bed file
}
